requirements.txt
# Core dependencies
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.25.0
google-generativeai>=0.3.0
pyyaml>=6.0
numpy>=1.24.0

# Azure OpenAI
openai>=1.0.0

# LangGraph for workflow orchestration
langgraph>=0.0.20

# NLP and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
nltk>=3.8.0
faiss-cpu>=1.7.0  # For duplicate detection with FAISS

# CLI and utilities
click>=8.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Quality metrics
textstat>=0.7.3

# Visualization and sentiment analysis
matplotlib>=3.7.0
seaborn>=0.12.0
vaderSentiment>=3.3.2

datasets == 4.5.0
